Top level cleaned up
